Command Modularization
Check CHANGELOG.MD for changes made
