+++ chat.py
@@ -1,6 +1,6 @@
 import google.generativeai as genai
 from config import API_KEY
-import config
+from command_modules_init import all_commands_str, additional_info_str
 
 generation_config = {
     "temperature": 1,
@@ -25,36 +25,12 @@ If you are asked to execute a command, you can still talk to Iron Man and confir
 you've taken or your thoughts on them.
 
 Commands:
-  prs(q)      - Nuke, destroy, or use maximum pulse  
-  prs(e)      - Power up  
-  prs(shift)  - Start or stop flying  
-  prs(f)      - Fire rockets  
-  prs(r)      - Reload  
-  fre(n)      - Fire a specific number of shots  
-  uni(n)      - Fire the unibeam for a specific number of seconds  
-  msg(txt true) - Send a message in team chat  
-  msg(txt false) - Send a message in match chat  
-  frk;        - Activate freaky mode  
-  clp;        - Save or clip replays  
-  str;        - Start recording  
-  spr;        - Stop recording  
-  dly(n)      - Delay execution by n seconds  
-  ext;        - Power off or terminate the program  
-  mut;        - Mute or unmute audio  
-  vld;        - Turn down the volume  
-  vlu;        - Turn up the volume  
-  vls(n)      - Set volume to a specific percentage  
-  lck;        - Insta-lock Iron Man  
-  ply(name)   - Play a sound effect (fart, roast, laugh, getout, boom, scream, yay)  
-  mpl(name)   - Start a music playlist (Playlist1 or Playlist2)  
-  mps;        - Pause or resume music  
-  skp;        - Skip the current song  
+"""
++ all_commands_str
++ additional_info_str
++
+"""
 
-Iron Man could also ask you to blame his teammates depending on their roles for losing:  
-  - dps for damage dealers  
-  - supports for healers  
-  - tanks for defenders  
-In that case you send a message in team chat addressing the issue, And you are encouraged to be toxic in this case unless instructed otherwise
 Remember to actually send the commands when requested, and send them EXACTLY as they're written, 
 with no missing `;` or misplaced `()`.
 """
@@ -62,14 +38,13 @@ with no missing `;` or misplaced `()`.
 
 chat_session = model.start_chat()
 
-def send_to_jarvis(message, process_command_func, speak_func, play_sound_func):
+def send_to_jarvis(message, process_command_func, speak_func):
     """
     Send the user's message to the AI model and process the response.
     The response text is parsed for commands (via process_command_func), then spoken and any sound effects played.
     """
     response = chat_session.send_message(message)
-    cleaned_response, sound_effects = process_command_func(response.text)
+    cleaned_response = process_command_func(response.text)
     print("Jarvis:", response.text)
     speak_func(cleaned_response)
-    for sound in sound_effects:
-        play_sound_func(sound)
+
