+++ tasks.py
@@ -1,6 +1,5 @@
 import threading
 import queue
-import time
 
 # A global task queue for scheduling actions.
 task_queue = queue.Queue()
@@ -10,12 +9,17 @@ def task_worker():
         func, args = task_queue.get()
         if func is None:
             break
-        func(*args)
+        
+        if (isinstance(args, tuple)):
+            func(*args)
+        else:
+            func(args)
+            
         task_queue.task_done()
 
 # Start a background worker thread.
 worker_thread = threading.Thread(target=task_worker, daemon=True)
 worker_thread.start()
 
-def add_task(func, *args):
+def add_task(func, args):
     task_queue.put((func, args))
